pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "sendgrid>=6.11.0",
    "twilio>=9.5.2",
    "pillow>=11.2.1",
    "reportlab>=4.4.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.1.2",
    "pypdf2>=3.0.1",
    "pyinstaller>=6.13.0",
    "beautifulsoup4>=4.13.4",
    "openai>=1.76.0",
    "anthropic>=0.50.0",
]
